This patch has two new variables #*CABPATH* and #*GUID* and two new options *cab_path* and *cab_path_autodelete*. TEMPPATH is now back to being just TEMPPATH, so if you are CABing, you should be using CABBPATH instead of TEMPPATH to run dependencies. Copied from https://dotnetinstaller.svn.sourceforge.net/svnroot/dotnetinstaller/trunk/, rev. 29 by dblockdotorg @ 6/12/2008 10:34 PM

+++ dotNetInstaller/dotNetInstallerDlg.cpp
@@ -395,7 +395,7 @@ void CdotNetInstallerDlg::OnDestroy()
         // 2008-06-09 - Daniel Doubrovkine - delete temporary directory
         // even if a reboot is required, the temporary folder is gone; next run will re-extract components
         CString temppath = DVLib::GetSessionTempPath(true);
-        if (temppath.GetLength() && DVLib::FileExistsCustom(temppath))
+        if (m_Settings.cab_path_autodelete && temppath.GetLength() && DVLib::FileExistsCustom(temppath))
         {
 		    ApplicationLog.Write(TEXT("Deleting temporary folder: ") + temppath);
             DVLib::DeleteDirectoryDeep(temppath);
@@ -414,7 +414,7 @@ void CdotNetInstallerDlg::OnBnClickedCancel()
 
 void CdotNetInstallerDlg::ExtractCab()
 {
-    ExtractCABComponent e_component;
+    ExtractCABComponent e_component(m_Settings);
     installerSetting e_setting;
     e_setting.installing_component_wait = m_Settings.cab_dialog_message;
     e_component.description = m_Settings.cab_dialog_caption;
